+++ database/migrations/2022_04_09_121924_create_purchase_transactions_table.php
@@ -47,7 +47,7 @@ class CreatePurchaseTransactionsTable extends Migration
 
             $table->dateTime('finalized_at');
 
-            $table->integer('verification_status')->unsigned()->default(2);
+            $table->tinyInteger('verification_status')->unsigned()->default(2);
 
             $table->string('verification_note')->nullable();
